% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{circle2d}
\alias{circle2d}
\title{2-dimensional point cloud of a unit circle}
\format{A matrix with 100 rows and 2 columns: the x- and y-coordinates}
\source{
\url{https://github.com/rrrlw/TDAstats/blob/master/data-raw/circle2d.R}
}
\usage{
circle2d
}
\description{
A dataset containing the Cartesian coordinates of 100 points uniformly distributed
on the circumference of a unit circle.
}
\keyword{datasets}
