% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sphere3d}
\alias{sphere3d}
\title{3-dimensional point cloud of a unit sphere}
\format{A matrix with 100 rows and 3 columns: the x-, y-, and z-coordinates}
\source{
\url{https://github.com/rrrlw/TDAstats/blob/master/data-raw/sphere3d.R}
}
\usage{
sphere3d
}
\description{
A dataset containing the Cartesian coordinates of 100 points uniformly distributed
on the surface of a unit sphere.
}
\keyword{datasets}
