% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unif2d}
\alias{unif2d}
\title{2-dimensional point cloud of a unit square}
\format{A matrix with 100 rows and 2 columns: the x- and y-coordinates}
\source{
\url{https://github.com/rrrlw/TDAstats/blob/master/data-raw/unif2d.R}
}
\usage{
unif2d
}
\description{
A dataset containing the Cartesian coordinates of 100 points uniformly distributed
within a unit square.
}
\keyword{datasets}
