% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unif3d}
\alias{unif3d}
\title{3-dimensional point cloud of a unit cube}
\format{A matrix with 100 rows and 3 columns: the x-, y-, and z-coordinates}
\source{
\url{https://github.com/rrrlw/TDAstats/blob/master/data-raw/unif3d.R}
}
\usage{
unif3d
}
\description{
A dataset containing the Cartesian coordinates of 100 points uniformly distributed
within a unit cube.
}
\keyword{datasets}
