% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDAstats.R
\docType{package}
\name{TDAstats}
\alias{TDAstats}
\alias{TDAstats-package}
\title{Statistical Inference for Persistent Homology in Topological Data Analysis}
\description{
This package aims to be a comprehensive toolset for any
useR conducting topological data analysis, specifically via the
calculation of persistent homology in a Vietoris-Rips complex.
The tools this package currently provides can be conveniently split
into three main sections: (1) calculating persistent homology; (2)
conducting statistical inference on persistent homology calculations;
(3) visualizing persistent homology and statistical inference.
}
